Extract numeric normalization utility
